requirements/prod.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.0
django-unfold==0.67.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
Faker==23.3.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
